[ALLUXIO-2743] Merge ConstantPropertyKey to PropertyKey
