--- core/common/src/main/java/alluxio/ConstantPropertyKey.java
@@ -1,510 +1,0 @@
-/*
- * The Alluxio Open Foundation licenses this work under the Apache License, version 2.0
- * (the "License"). You may not use this work except in compliance with the License, which is
- * available at www.apache.org/licenses/LICENSE-2.0
- *
- * This software is distributed on an "AS IS" basis, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
- * either express or implied, as more fully set forth in the License.
- *
- * See the NOTICE file distributed with this work for information regarding copyright ownership.
- */
-
-package alluxio;
-
-import java.util.Collection;
-import java.util.HashMap;
-import java.util.Map;
-
-import javax.annotation.concurrent.ThreadSafe;
-
-/**
- * A property key that is predefined at compile time.
- */
-@ThreadSafe
-public final class ConstantPropertyKey extends PropertyKey {
-  /**
-   * A nested class to hold named string constants for their corresponding properties.
-   * Used for setting configuration in integration tests.
-   */
-  @ThreadSafe
-  public static final class Name {
-    public static final String CONF_DIR = "alluxio.conf.dir";
-    public static final String DEBUG = "alluxio.debug";
-    public static final String HOME = "alluxio.home";
-    public static final String INTEGRATION_MASTER_RESOURCE_CPU =
-        "alluxio.integration.master.resource.cpu";
-    public static final String INTEGRATION_MASTER_RESOURCE_MEM =
-        "alluxio.integration.master.resource.mem";
-    public static final String INTEGRATION_MESOS_ALLUXIO_MASTER_NAME =
-        "alluxio.integration.mesos.master.name";
-    public static final String INTEGRATION_MESOS_ALLUXIO_MASTER_NODE_COUNT =
-        "alluxio.integration.mesos.master.node.count";
-    public static final String INTEGRATION_MESOS_ALLUXIO_WORKER_NAME =
-        "alluxio.integration.mesos.worker.name";
-    public static final String INTEGRATION_MESOS_ALLUXIO_JAR_URL =
-        "alluxio.integration.mesos.alluxio.jar.url";
-    public static final String INTEGRATION_MESOS_JDK_PATH = "alluxio.integration.mesos.jdk.path";
-    public static final String INTEGRATION_MESOS_JDK_URL = "alluxio.integration.mesos.jdk.url";
-    public static final String INTEGRATION_MESOS_PRINCIPAL = "alluxio.integration.mesos.principal";
-    public static final String INTEGRATION_MESOS_ROLE = "alluxio.integration.mesos.role";
-    public static final String INTEGRATION_MESOS_SECRET = "alluxio.integration.mesos.secret";
-    public static final String INTEGRATION_MESOS_USER = "alluxio.integration.mesos.user";
-    public static final String INTEGRATION_WORKER_RESOURCE_CPU =
-        "alluxio.integration.worker.resource.cpu";
-    public static final String INTEGRATION_WORKER_RESOURCE_MEM =
-        "alluxio.integration.worker.resource.mem";
-    public static final String INTEGRATION_YARN_WORKERS_PER_HOST_MAX =
-        "alluxio.integration.yarn.workers.per.host.max";
-    public static final String KEY_VALUE_ENABLED = "alluxio.keyvalue.enabled";
-    public static final String KEY_VALUE_PARTITION_SIZE_BYTES_MAX =
-        "alluxio.keyvalue.partition.size.bytes.max";
-    public static final String LOGGER_TYPE = "alluxio.logger.type";
-    public static final String LOGS_DIR = "alluxio.logs.dir";
-    public static final String METRICS_CONF_FILE = "alluxio.metrics.conf.file";
-    public static final String NETWORK_HOST_RESOLUTION_TIMEOUT_MS =
-        "alluxio.network.host.resolution.timeout.ms";
-    public static final String NETWORK_THRIFT_FRAME_SIZE_BYTES_MAX =
-        "alluxio.network.thrift.frame.size.bytes.max";
-    public static final String SITE_CONF_DIR = "alluxio.site.conf.dir";
-    public static final String TEST_MODE = "alluxio.test.mode";
-    public static final String VERSION = "alluxio.version";
-    public static final String WEB_RESOURCES = "alluxio.web.resources";
-    public static final String WEB_THREADS = "alluxio.web.threads";
-    public static final String WORK_DIR = "alluxio.work.dir";
-    public static final String ZOOKEEPER_ADDRESS = "alluxio.zookeeper.address";
-    public static final String ZOOKEEPER_ELECTION_PATH = "alluxio.zookeeper.election.path";
-    public static final String ZOOKEEPER_ENABLED = "alluxio.zookeeper.enabled";
-    public static final String ZOOKEEPER_LEADER_INQUIRY_RETRY_COUNT =
-        "alluxio.zookeeper.leader.inquiry.retry";
-    public static final String ZOOKEEPER_LEADER_PATH = "alluxio.zookeeper.leader.path";
-
-    //
-    // UFS related properties
-    //
-    public static final String UNDERFS_ADDRESS = "alluxio.underfs.address";
-    public static final String UNDERFS_ALLOW_SET_OWNER_FAILURE =
-        "alluxio.underfs.allow.set.owner.failure";
-    public static final String UNDERFS_LISTING_LENGTH = "alluxio.underfs.listing.length";
-    public static final String UNDERFS_GCS_OWNER_ID_TO_USERNAME_MAPPING =
-        "alluxio.underfs.gcs.owner.id.to.username.mapping";
-    public static final String UNDERFS_GLUSTERFS_IMPL = "alluxio.underfs.glusterfs.impl";
-    public static final String UNDERFS_GLUSTERFS_MOUNTS = "alluxio.underfs.glusterfs.mounts";
-    public static final String UNDERFS_GLUSTERFS_MR_DIR =
-        "alluxio.underfs.glusterfs.mapred.system.dir";
-    public static final String UNDERFS_GLUSTERFS_VOLUMES = "alluxio.underfs.glusterfs.volumes";
-    public static final String UNDERFS_HDFS_CONFIGURATION = "alluxio.underfs.hdfs.configuration";
-    public static final String UNDERFS_HDFS_IMPL = "alluxio.underfs.hdfs.impl";
-    public static final String UNDERFS_HDFS_PREFIXES = "alluxio.underfs.hdfs.prefixes";
-    public static final String UNDERFS_HDFS_REMOTE = "alluxio.underfs.hdfs.remote";
-    public static final String UNDERFS_OBJECT_STORE_MOUNT_SHARED_PUBLICLY =
-        "alluxio.underfs.object.store.mount.shared.publicly";
-    public static final String UNDERFS_OSS_CONNECT_MAX = "alluxio.underfs.oss.connection.max";
-    public static final String UNDERFS_OSS_CONNECT_TIMEOUT =
-        "alluxio.underfs.oss.connection.timeout.ms";
-    public static final String UNDERFS_OSS_CONNECT_TTL = "alluxio.underfs.oss.connection.ttl";
-    public static final String UNDERFS_OSS_SOCKET_TIMEOUT = "alluxio.underfs.oss.socket.timeout.ms";
-    public static final String UNDERFS_S3A_INHERIT_ACL = "alluxio.underfs.s3a.inherit_acl";
-    public static final String UNDERFS_S3A_CONSISTENCY_TIMEOUT_MS =
-        "alluxio.underfs.s3a.consistency.timeout.ms";
-    public static final String UNDERFS_S3A_REQUEST_TIMEOUT_MS =
-        "alluxio.underfs.s3a.request.timeout.ms";
-    public static final String UNDERFS_S3A_SECURE_HTTP_ENABLED =
-        "alluxio.underfs.s3a.secure.http.enabled";
-    public static final String UNDERFS_S3A_SERVER_SIDE_ENCRYPTION_ENABLED =
-        "alluxio.underfs.s3a.server.side.encryption.enabled";
-    public static final String UNDERFS_S3A_SOCKET_TIMEOUT_MS =
-        "alluxio.underfs.s3a.socket.timeout.ms";
-    public static final String UNDERFS_S3_ADMIN_THREADS_MAX =
-        "alluxio.underfs.s3.admin.threads.max";
-    public static final String UNDERFS_S3_DISABLE_DNS_BUCKETS =
-        "alluxio.underfs.s3.disable.dns.buckets";
-    public static final String UNDERFS_S3_ENDPOINT = "alluxio.underfs.s3.endpoint";
-    public static final String UNDERFS_S3_ENDPOINT_HTTPS_PORT =
-        "alluxio.underfs.s3.endpoint.https.port";
-    public static final String UNDERFS_S3_ENDPOINT_HTTP_PORT =
-        "alluxio.underfs.s3.endpoint.http.port";
-    public static final String UNDERFS_S3_OWNER_ID_TO_USERNAME_MAPPING =
-        "alluxio.underfs.s3.owner.id.to.username.mapping";
-    public static final String UNDERFS_S3_PROXY_HOST = "alluxio.underfs.s3.proxy.host";
-    public static final String UNDERFS_S3_PROXY_HTTPS_ONLY = "alluxio.underfs.s3.proxy.https.only";
-    public static final String UNDERFS_S3_PROXY_PORT = "alluxio.underfs.s3.proxy.port";
-    public static final String UNDERFS_S3_THREADS_MAX = "alluxio.underfs.s3.threads.max";
-    public static final String UNDERFS_S3_UPLOAD_THREADS_MAX =
-        "alluxio.underfs.s3.upload.threads.max";
-
-    //
-    // UFS access control related properties
-    //
-    public static final String GCS_ACCESS_KEY = "fs.gcs.accessKeyId";
-    public static final String GCS_SECRET_KEY = "fs.gcs.secretAccessKey";
-    public static final String OSS_ACCESS_KEY = "fs.oss.accessKeyId";
-    public static final String OSS_ENDPOINT_KEY = "fs.oss.endpoint";
-    public static final String OSS_SECRET_KEY = "fs.oss.accessKeySecret";
-    public static final String S3A_ACCESS_KEY = "aws.accessKeyId";
-    public static final String S3A_SECRET_KEY = "aws.secretKey";
-    public static final String S3N_ACCESS_KEY = "fs.s3n.awsAccessKeyId";
-    public static final String S3N_SECRET_KEY = "fs.s3n.awsSecretAccessKey";
-    public static final String SWIFT_API_KEY = "fs.swift.apikey";
-    public static final String SWIFT_AUTH_METHOD_KEY = "fs.swift.auth.method";
-    public static final String SWIFT_AUTH_URL_KEY = "fs.swift.auth.url";
-    public static final String SWIFT_PASSWORD_KEY = "fs.swift.password";
-    public static final String SWIFT_SIMULATION = "fs.swift.simulation";
-    public static final String SWIFT_TENANT_KEY = "fs.swift.tenant";
-    public static final String SWIFT_USER_KEY = "fs.swift.user";
-    public static final String SWIFT_USE_PUBLIC_URI_KEY = "fs.swift.use.public.url";
-    public static final String SWIFT_REGION_KEY = "fs.swift.region";
-
-    //
-    // Master related properties
-    //
-    public static final String MASTER_ADDRESS = "alluxio.master.address";
-    public static final String MASTER_BIND_HOST = "alluxio.master.bind.host";
-    public static final String MASTER_CONNECTION_TIMEOUT_MS =
-        "alluxio.master.connection.timeout.ms";
-    public static final String MASTER_FILE_ASYNC_PERSIST_HANDLER =
-        "alluxio.master.file.async.persist.handler";
-    public static final String MASTER_FORMAT_FILE_PREFIX = "alluxio.master.format.file_prefix";
-    public static final String MASTER_HEARTBEAT_INTERVAL_MS =
-        "alluxio.master.heartbeat.interval.ms";
-    public static final String MASTER_HOSTNAME = "alluxio.master.hostname";
-    public static final String MASTER_JOURNAL_FLUSH_BATCH_TIME_MS =
-        "alluxio.master.journal.flush.batch.time.ms";
-    public static final String MASTER_JOURNAL_FLUSH_TIMEOUT_MS =
-        "alluxio.master.journal.flush.timeout.ms";
-    public static final String MASTER_JOURNAL_FOLDER = "alluxio.master.journal.folder";
-    public static final String MASTER_JOURNAL_FORMATTER_CLASS =
-        "alluxio.master.journal.formatter.class";
-    public static final String MASTER_JOURNAL_LOG_SIZE_BYTES_MAX =
-        "alluxio.master.journal.log.size.bytes.max";
-    public static final String MASTER_JOURNAL_TAILER_SHUTDOWN_QUIET_WAIT_TIME_MS =
-        "alluxio.master.journal.tailer.shutdown.quiet.wait.time.ms";
-    public static final String MASTER_JOURNAL_TAILER_SLEEP_TIME_MS =
-        "alluxio.master.journal.tailer.sleep.time.ms";
-    public static final String MASTER_KEYTAB_KEY_FILE = "alluxio.master.keytab.file";
-    public static final String MASTER_LINEAGE_CHECKPOINT_CLASS =
-        "alluxio.master.lineage.checkpoint.class";
-    public static final String MASTER_LINEAGE_CHECKPOINT_INTERVAL_MS =
-        "alluxio.master.lineage.checkpoint.interval.ms";
-    public static final String MASTER_LINEAGE_RECOMPUTE_INTERVAL_MS =
-        "alluxio.master.lineage.recompute.interval.ms";
-    public static final String MASTER_LINEAGE_RECOMPUTE_LOG_PATH =
-        "alluxio.master.lineage.recompute.log.path";
-    public static final String MASTER_PRINCIPAL = "alluxio.master.principal";
-    public static final String MASTER_RETRY = "alluxio.master.retry";
-    public static final String MASTER_RPC_PORT = "alluxio.master.port";
-    public static final String MASTER_STARTUP_CONSISTENCY_CHECK_ENABLED =
-        "alluxio.master.startup.consistency.check.enabled";
-    public static final String MASTER_TIERED_STORE_GLOBAL_LEVEL0_ALIAS =
-        "alluxio.master.tieredstore.global.level0.alias";
-    public static final String MASTER_TIERED_STORE_GLOBAL_LEVEL1_ALIAS =
-        "alluxio.master.tieredstore.global.level1.alias";
-    public static final String MASTER_TIERED_STORE_GLOBAL_LEVEL2_ALIAS =
-        "alluxio.master.tieredstore.global.level2.alias";
-    public static final String MASTER_TIERED_STORE_GLOBAL_LEVELS =
-        "alluxio.master.tieredstore.global.levels";
-    public static final String MASTER_TTL_CHECKER_INTERVAL_MS =
-        "alluxio.master.ttl.checker.interval.ms";
-    public static final String MASTER_WEB_BIND_HOST = "alluxio.master.web.bind.host";
-    public static final String MASTER_WEB_HOSTNAME = "alluxio.master.web.hostname";
-    public static final String MASTER_WEB_PORT = "alluxio.master.web.port";
-    public static final String MASTER_WHITELIST = "alluxio.master.whitelist";
-    public static final String MASTER_WORKER_THREADS_MAX = "alluxio.master.worker.threads.max";
-    public static final String MASTER_WORKER_THREADS_MIN = "alluxio.master.worker.threads.min";
-    public static final String MASTER_WORKER_TIMEOUT_MS = "alluxio.master.worker.timeout.ms";
-
-    //
-    // Worker related properties
-    //
-    public static final String WORKER_ALLOCATOR_CLASS = "alluxio.worker.allocator.class";
-    public static final String WORKER_BIND_HOST = "alluxio.worker.bind.host";
-    public static final String WORKER_BLOCK_HEARTBEAT_INTERVAL_MS =
-        "alluxio.worker.block.heartbeat.interval.ms";
-    public static final String WORKER_BLOCK_HEARTBEAT_TIMEOUT_MS =
-        "alluxio.worker.block.heartbeat.timeout.ms";
-    public static final String WORKER_BLOCK_THREADS_MAX = "alluxio.worker.block.threads.max";
-    public static final String WORKER_BLOCK_THREADS_MIN = "alluxio.worker.block.threads.min";
-    public static final String WORKER_DATA_BIND_HOST = "alluxio.worker.data.bind.host";
-    public static final String WORKER_DATA_FOLDER = "alluxio.worker.data.folder";
-    public static final String WORKER_DATA_HOSTNAME = "alluxio.worker.data.hostname";
-    public static final String WORKER_DATA_PORT = "alluxio.worker.data.port";
-    public static final String WORKER_DATA_SERVER_CLASS = "alluxio.worker.data.server.class";
-    public static final String WORKER_DATA_TMP_FOLDER = "alluxio.worker.data.folder.tmp";
-    public static final String WORKER_DATA_TMP_SUBDIR_MAX = "alluxio.worker.data.tmp.subdir.max";
-    public static final String WORKER_EVICTOR_CLASS = "alluxio.worker.evictor.class";
-    public static final String WORKER_EVICTOR_LRFU_ATTENUATION_FACTOR =
-        "alluxio.worker.evictor.lrfu.attenuation.factor";
-    public static final String WORKER_EVICTOR_LRFU_STEP_FACTOR =
-        "alluxio.worker.evictor.lrfu.step.factor";
-    public static final String WORKER_FILESYSTEM_HEARTBEAT_INTERVAL_MS =
-        "alluxio.worker.filesystem.heartbeat.interval.ms";
-    public static final String WORKER_FILE_PERSIST_POOL_SIZE =
-        "alluxio.worker.file.persist.pool.size";
-    public static final String WORKER_FILE_PERSIST_RATE_LIMIT =
-        "alluxio.worker.file.persist.rate.limit";
-    public static final String WORKER_FILE_PERSIST_RATE_LIMIT_ENABLED =
-        "alluxio.worker.file.persist.rate.limit.enabled";
-    public static final String WORKER_FILE_BUFFER_SIZE = "alluxio.worker.file.buffer.size";
-    public static final String WORKER_HOSTNAME = "alluxio.worker.hostname";
-    public static final String WORKER_KEYTAB_FILE = "alluxio.worker.keytab.file";
-    public static final String WORKER_MEMORY_SIZE = "alluxio.worker.memory.size";
-    public static final String WORKER_NETWORK_NETTY_BACKLOG =
-        "alluxio.worker.network.netty.backlog";
-    public static final String WORKER_NETWORK_NETTY_BOSS_THREADS =
-        "alluxio.worker.network.netty.boss.threads";
-    public static final String WORKER_NETWORK_NETTY_BUFFER_RECEIVE =
-        "alluxio.worker.network.netty.buffer.receive";
-    public static final String WORKER_NETWORK_NETTY_BUFFER_SEND =
-        "alluxio.worker.network.netty.buffer.send";
-    public static final String WORKER_NETWORK_NETTY_CHANNEL =
-        "alluxio.worker.network.netty.channel";
-    public static final String WORKER_NETWORK_NETTY_FILE_TRANSFER_TYPE =
-        "alluxio.worker.network.netty.file.transfer";
-    public static final String WORKER_NETWORK_NETTY_SHUTDOWN_QUIET_PERIOD =
-        "alluxio.worker.network.netty.shutdown.quiet.period";
-    public static final String WORKER_NETWORK_NETTY_SHUTDOWN_TIMEOUT =
-        "alluxio.worker.network.netty.shutdown.timeout";
-    public static final String WORKER_NETWORK_NETTY_WATERMARK_HIGH =
-        "alluxio.worker.network.netty.watermark.high";
-    public static final String WORKER_NETWORK_NETTY_WATERMARK_LOW =
-        "alluxio.worker.network.netty.watermark.low";
-    public static final String WORKER_NETWORK_NETTY_WORKER_THREADS =
-        "alluxio.worker.network.netty.worker.threads";
-    public static final String WORKER_NETWORK_NETTY_WRITER_BUFFER_SIZE_PACKETS =
-        "alluxio.worker.network.netty.writer.buffer.size.packets";
-    public static final String WORKER_NETWORK_NETTY_READER_BUFFER_SIZE_PACKETS =
-        "alluxio.worker.network.netty.reader.buffer.size.packets";
-    public static final String WORKER_NETWORK_NETTY_READER_PACKET_SIZE_BYTES =
-        "alluxio.worker.network.netty.reader.packet.size.bytes";
-    public static final String WORKER_NETWORK_NETTY_BLOCK_READER_THREADS_MAX =
-        "alluxio.worker.network.netty.block.reader.threads.max";
-    public static final String WORKER_NETWORK_NETTY_BLOCK_WRITER_THREADS_MAX =
-        "alluxio.worker.network.netty.block.writer.threads.max";
-    public static final String WORKER_NETWORK_NETTY_FILE_READER_THREADS_MAX =
-        "alluxio.worker.network.netty.file.reader.threads.max";
-    public static final String WORKER_NETWORK_NETTY_FILE_WRITER_THREADS_MAX =
-        "alluxio.worker.network.netty.file.writer.threads.max";
-    public static final String WORKER_PRINCIPAL = "alluxio.worker.principal";
-    public static final String WORKER_RPC_PORT = "alluxio.worker.port";
-    public static final String WORKER_SESSION_TIMEOUT_MS = "alluxio.worker.session.timeout.ms";
-    public static final String WORKER_TIERED_STORE_BLOCK_LOCK_READERS =
-        "alluxio.worker.tieredstore.block.lock.readers";
-    public static final String WORKER_TIERED_STORE_BLOCK_LOCKS =
-        "alluxio.worker.tieredstore.block.locks";
-    public static final String WORKER_TIERED_STORE_LEVEL0_ALIAS =
-        "alluxio.worker.tieredstore.level0.alias";
-    public static final String WORKER_TIERED_STORE_LEVEL0_DIRS_PATH =
-        "alluxio.worker.tieredstore.level0.dirs.path";
-    public static final String WORKER_TIERED_STORE_LEVEL0_DIRS_QUOTA =
-        "alluxio.worker.tieredstore.level0.dirs.quota";
-    public static final String WORKER_TIERED_STORE_LEVEL0_RESERVED_RATIO =
-        "alluxio.worker.tieredstore.level0.reserved.ratio";
-    public static final String WORKER_TIERED_STORE_LEVEL1_ALIAS =
-        "alluxio.worker.tieredstore.level1.alias";
-    public static final String WORKER_TIERED_STORE_LEVEL1_DIRS_PATH =
-        "alluxio.worker.tieredstore.level1.dirs.path";
-    public static final String WORKER_TIERED_STORE_LEVEL1_DIRS_QUOTA =
-        "alluxio.worker.tieredstore.level1.dirs.quota";
-    public static final String WORKER_TIERED_STORE_LEVEL1_RESERVED_RATIO =
-        "alluxio.worker.tieredstore.level1.reserved.ratio";
-    public static final String WORKER_TIERED_STORE_LEVEL2_ALIAS =
-        "alluxio.worker.tieredstore.level2.alias";
-    public static final String WORKER_TIERED_STORE_LEVEL2_DIRS_PATH =
-        "alluxio.worker.tieredstore.level2.dirs.path";
-    public static final String WORKER_TIERED_STORE_LEVEL2_DIRS_QUOTA =
-        "alluxio.worker.tieredstore.level2.dirs.quota";
-    public static final String WORKER_TIERED_STORE_LEVEL2_RESERVED_RATIO =
-        "alluxio.worker.tieredstore.level2.reserved.ratio";
-    public static final String WORKER_TIERED_STORE_LEVELS = "alluxio.worker.tieredstore.levels";
-    public static final String WORKER_TIERED_STORE_RESERVER_ENABLED =
-        "alluxio.worker.tieredstore.reserver.enabled";
-    public static final String WORKER_TIERED_STORE_RESERVER_INTERVAL_MS =
-        "alluxio.worker.tieredstore.reserver.interval.ms";
-    public static final String WORKER_TIERED_STORE_RETRY = "alluxio.worker.tieredstore.retry";
-    public static final String WORKER_WEB_BIND_HOST = "alluxio.worker.web.bind.host";
-    public static final String WORKER_WEB_HOSTNAME = "alluxio.worker.web.hostname";
-    public static final String WORKER_WEB_PORT = "alluxio.worker.web.port";
-
-    //
-    // Proxy related properties
-    //
-    public static final String PROXY_STREAM_CACHE_TIMEOUT_MS =
-        "alluxio.proxy.stream.cache.timeout.ms";
-    public static final String PROXY_WEB_BIND_HOST = "alluxio.proxy.web.bind.host";
-    public static final String PROXY_WEB_HOSTNAME = "alluxio.proxy.web.hostname";
-    public static final String PROXY_WEB_PORT = "alluxio.proxy.web.port";
-
-    //
-    // User related properties
-    //
-    public static final String USER_BLOCK_MASTER_CLIENT_THREADS =
-        "alluxio.user.block.master.client.threads";
-    public static final String USER_BLOCK_REMOTE_READER_CLASS =
-        "alluxio.user.block.remote.reader.class";
-    public static final String USER_BLOCK_REMOTE_READ_BUFFER_SIZE_BYTES =
-        "alluxio.user.block.remote.read.buffer.size.bytes";
-    public static final String USER_BLOCK_REMOTE_WRITER_CLASS =
-        "alluxio.user.block.remote.writer.class";
-    public static final String USER_BLOCK_SIZE_BYTES_DEFAULT =
-        "alluxio.user.block.size.bytes.default";
-    public static final String USER_BLOCK_WORKER_CLIENT_THREADS =
-        "alluxio.user.block.worker.client.threads";
-    public static final String USER_BLOCK_WORKER_CLIENT_POOL_SIZE_MAX =
-        "alluxio.user.block.worker.client.pool.size.max";
-    public static final String USER_BLOCK_WORKER_CLIENT_POOL_GC_THRESHOLD_MS =
-        "alluxio.user.block.worker.client.pool.gc.threshold.ms";
-    public static final String USER_DATE_FORMAT_PATTERN = "alluxio.user.date.format.pattern";
-    public static final String USER_FAILED_SPACE_REQUEST_LIMITS =
-        "alluxio.user.failed.space.request.limits";
-    public static final String USER_FILE_BUFFER_BYTES = "alluxio.user.file.buffer.bytes";
-    public static final String USER_FILE_CACHE_PARTIALLY_READ_BLOCK =
-        "alluxio.user.file.cache.partially.read.block";
-    public static final String USER_FILE_MASTER_CLIENT_THREADS =
-        "alluxio.user.file.master.client.threads";
-    public static final String USER_FILE_PASSIVE_CACHE_ENABLED =
-        "alluxio.user.file.passive.cache.enabled";
-    public static final String USER_FILE_READ_TYPE_DEFAULT = "alluxio.user.file.readtype.default";
-    public static final String USER_FILE_SEEK_BUFFER_SIZE_BYTES =
-        "alluxio.user.file.seek.buffer.size.bytes";
-    public static final String USER_FILE_WAITCOMPLETED_POLL_MS =
-        "alluxio.user.file.waitcompleted.poll.ms";
-    public static final String USER_FILE_WORKER_CLIENT_THREADS =
-        "alluxio.user.file.worker.client.threads";
-    public static final String USER_FILE_WORKER_CLIENT_POOL_SIZE_MAX =
-        "alluxio.user.file.worker.client.pool.size.max";
-    public static final String USER_FILE_WORKER_CLIENT_POOL_GC_THRESHOLD_MS =
-        "alluxio.user.file.worker.client.pool.gc.threshold.ms";
-    public static final String USER_FILE_WRITE_LOCATION_POLICY =
-        "alluxio.user.file.write.location.policy.class";
-    public static final String USER_FILE_WRITE_AVOID_EVICTION_POLICY_RESERVED_BYTES =
-        "alluxio.user.file.write.avoid.eviction.policy.reserved.size.bytes";
-    public static final String USER_FILE_WRITE_TYPE_DEFAULT = "alluxio.user.file.writetype.default";
-    public static final String USER_FILE_WRITE_TIER_DEFAULT =
-        "alluxio.user.file.write.tier.default";
-    public static final String USER_HEARTBEAT_INTERVAL_MS = "alluxio.user.heartbeat.interval.ms";
-    public static final String USER_HOSTNAME = "alluxio.user.hostname";
-    public static final String USER_LINEAGE_ENABLED = "alluxio.user.lineage.enabled";
-    public static final String USER_LINEAGE_MASTER_CLIENT_THREADS =
-        "alluxio.user.lineage.master.client.threads";
-    public static final String USER_LOCAL_READER_PACKET_SIZE_BYTES =
-        "alluxio.user.local.reader.packet.size.bytes";
-    public static final String USER_LOCAL_WRITER_PACKET_SIZE_BYTES =
-        "alluxio.user.local.writer.packet.size.bytes";
-    public static final String USER_NETWORK_NETTY_CHANNEL = "alluxio.user.network.netty.channel";
-    public static final String USER_NETWORK_NETTY_TIMEOUT_MS =
-        "alluxio.user.network.netty.timeout.ms";
-    public static final String USER_NETWORK_NETTY_WORKER_THREADS =
-        "alluxio.user.network.netty.worker.threads";
-    public static final String USER_NETWORK_NETTY_CHANNEL_POOL_SIZE_MAX =
-        "alluxio.user.network.netty.channel.pool.size.max";
-    public static final String USER_NETWORK_NETTY_CHANNEL_POOL_GC_THRESHOLD_MS =
-        "alluxio.user.network.netty.channel.pool.gc.threshold.ms";
-    public static final String USER_NETWORK_NETTY_CHANNEL_POOL_DISABLED =
-        "alluxio.user.network.netty.channel.pool.disabled";
-    public static final String USER_NETWORK_NETTY_WRITER_PACKET_SIZE_BYTES =
-        "alluxio.user.network.netty.writer.packet.size.bytes";
-    public static final String USER_NETWORK_NETTY_WRITER_BUFFER_SIZE_PACKETS =
-        "alluxio.user.network.netty.writer.buffer.size.packets";
-    public static final String USER_NETWORK_NETTY_READER_BUFFER_SIZE_PACKETS =
-        "alluxio.user.network.netty.reader.buffer.size.packets";
-    public static final String USER_NETWORK_NETTY_READER_CANCEL_ENABLED =
-        "alluxio.user.network.netty.reader.cancel.enabled";
-    public static final String USER_PACKET_STREAMING_ENABLED =
-        "alluxio.user.packet.streaming.enabled";
-    public static final String USER_RPC_RETRY_BASE_SLEEP_MS =
-        "alluxio.user.rpc.retry.base.sleep.ms";
-    public static final String USER_RPC_RETRY_MAX_NUM_RETRY =
-        "alluxio.user.rpc.retry.max.num.retry";
-    public static final String USER_RPC_RETRY_MAX_SLEEP_MS = "alluxio.user.rpc.retry.max.sleep.ms";
-    public static final String USER_UFS_DELEGATION_ENABLED = "alluxio.user.ufs.delegation.enabled";
-    public static final String USER_UFS_DELEGATION_READ_BUFFER_SIZE_BYTES =
-        "alluxio.user.ufs.delegation.read.buffer.size.bytes";
-    public static final String USER_UFS_DELEGATION_WRITE_BUFFER_SIZE_BYTES =
-        "alluxio.user.ufs.delegation.write.buffer.size.bytes";
-    public static final String USER_UFS_FILE_READER_CLASS = "alluxio.user.ufs.file.reader.class";
-    public static final String USER_UFS_FILE_WRITER_CLASS = "alluxio.user.ufs.file.writer.class";
-    public static final String USER_UFS_BLOCK_READ_LOCATION_POLICY =
-        "alluxio.user.ufs.block.read.location.policy";
-    public static final String USER_UFS_BLOCK_READ_LOCATION_POLICY_DETERMINISTIC_HASH_SHARDS =
-        "alluxio.user.ufs.block.read.location.policy.deterministic.hash.shards";
-    public static final String USER_UFS_BLOCK_READ_CONCURRENCY_MAX =
-        "alluxio.user.ufs.block.read.concurrency.max";
-    public static final String USER_UFS_BLOCK_OPEN_TIMEOUT_MS =
-        "alluxio.user.ufs.block.open.timeout.ms";
-    public static final String USER_SHORT_CIRCUIT_ENABLED = "alluxio.user.short.circuit.disabled";
-
-    //
-    // FUSE integration related properties
-    //
-    public static final String FUSE_CACHED_PATHS_MAX = "alluxio.fuse.cached.paths.max";
-    public static final String FUSE_DEBUG_ENABLED = "alluxio.fuse.debug.enabled";
-    public static final String FUSE_FS_NAME = "alluxio.fuse.fs.name";
-    public static final String FUSE_FS_ROOT = "alluxio.fuse.fs.root";
-    public static final String FUSE_MAXWRITE_BYTES = "alluxio.fuse.maxwrite.bytes";
-    public static final String FUSE_MOUNT_DEFAULT = "alluxio.fuse.mount.default";
-
-    //
-    // Security related properties
-    //
-    public static final String SECURITY_AUTHENTICATION_CUSTOM_PROVIDER_CLASS =
-        "alluxio.security.authentication.custom.provider.class";
-    public static final String SECURITY_AUTHENTICATION_SOCKET_TIMEOUT_MS =
-        "alluxio.security.authentication.socket.timeout.ms";
-    public static final String SECURITY_AUTHENTICATION_TYPE =
-        "alluxio.security.authentication.type";
-    public static final String SECURITY_AUTHORIZATION_PERMISSION_ENABLED =
-        "alluxio.security.authorization.permission.enabled";
-    public static final String SECURITY_AUTHORIZATION_PERMISSION_SUPERGROUP =
-        "alluxio.security.authorization.permission.supergroup";
-    public static final String SECURITY_AUTHORIZATION_PERMISSION_UMASK =
-        "alluxio.security.authorization.permission.umask";
-    public static final String SECURITY_GROUP_MAPPING_CACHE_TIMEOUT_MS =
-        "alluxio.security.group.mapping.cache.timeout.ms";
-    public static final String SECURITY_GROUP_MAPPING_CLASS =
-        "alluxio.security.group.mapping.class";
-    public static final String SECURITY_LOGIN_USERNAME = "alluxio.security.login.username";
-
-    private Name() {
-    } // prevent instantiation
-  }
-
-  /** A map from default property key's string name to the key. */
-  static final Map<String, ConstantPropertyKey> DEFAULT_KEYS_MAP = new HashMap<>();
-
-  /**
-   * @param keyStr string of property key
-   * @return whether the input is a valid property name
-   */
-  public static boolean isValid(String keyStr) {
-    return DEFAULT_KEYS_MAP.containsKey(keyStr);
-  }
-
-  /**
-   * Parses a string and return its corresponding {@link PropertyKey}, return null if no such
-   * a property can be found.
-   *
-   * @param keyStr string of property key
-   * @return corresponding property
-   */
-  public static PropertyKey fromString(String keyStr) {
-    ConstantPropertyKey key = DEFAULT_KEYS_MAP.get(keyStr);
-    if (key != null) {
-      return key;
-    }
-    return null;
-  }
-
-  /**
-   * @return all default keys
-   */
-  public static Collection<ConstantPropertyKey> getDefaultKeys() {
-    return DEFAULT_KEYS_MAP.values();
-  }
-
-  /**
-   * @param name String of this property
-   */
-  ConstantPropertyKey(String name) {
-    super(name);
-    DEFAULT_KEYS_MAP.put(name, this);
-  }
-}
