+++ tests/src/test/java/alluxio/master/lineage/LineageMasterClientRestApiTest.java
@@ -12,9 +12,9 @@
 package alluxio.master.lineage;
 
 import alluxio.AlluxioURI;
-import alluxio.ConstantPropertyKey;
 import alluxio.Constants;
 import alluxio.LocalAlluxioClusterResource.Config;
+import alluxio.PropertyKey;
 import alluxio.client.file.FileSystemContext;
 import alluxio.client.lineage.LineageContext;
 import alluxio.client.lineage.LineageFileSystem;
@@ -59,14 +59,14 @@ public final class LineageMasterClientRestApiTest extends RestApiTest {
   }
 
   @Test
-  @Config(confParams = {ConstantPropertyKey.Name.USER_LINEAGE_ENABLED, "true"})
+  @Config(confParams = {PropertyKey.Name.USER_LINEAGE_ENABLED, "true"})
   public void serviceName() throws Exception {
     new TestCase(mHostname, mPort, getEndpoint(LineageMasterClientRestServiceHandler.SERVICE_NAME),
         NO_PARAMS, HttpMethod.GET, Constants.LINEAGE_MASTER_CLIENT_SERVICE_NAME).run();
   }
 
   @Test
-  @Config(confParams = {ConstantPropertyKey.Name.USER_LINEAGE_ENABLED, "true"})
+  @Config(confParams = {PropertyKey.Name.USER_LINEAGE_ENABLED, "true"})
   public void serviceVersion() throws Exception {
     new TestCase(mHostname, mPort,
         getEndpoint(LineageMasterClientRestServiceHandler.SERVICE_VERSION), NO_PARAMS,
@@ -74,7 +74,7 @@ public final class LineageMasterClientRestApiTest extends RestApiTest {
   }
 
   @Test
-  @Config(confParams = {ConstantPropertyKey.Name.USER_LINEAGE_ENABLED, "true"})
+  @Config(confParams = {PropertyKey.Name.USER_LINEAGE_ENABLED, "true"})
   public void createLineage() throws Exception {
     mLineageClient.createFile(new AlluxioURI("/input")).close();
 
@@ -89,7 +89,7 @@ public final class LineageMasterClientRestApiTest extends RestApiTest {
   }
 
   @Test
-  @Config(confParams = {ConstantPropertyKey.Name.USER_LINEAGE_ENABLED, "true"})
+  @Config(confParams = {PropertyKey.Name.USER_LINEAGE_ENABLED, "true"})
   public void deleteLineage() throws Exception {
     LineageMaster lineageMaster = mMaster.getMaster(LineageMaster.class);
     long lineageId = lineageMaster.createLineage(new ArrayList<AlluxioURI>(),
@@ -104,7 +104,7 @@ public final class LineageMasterClientRestApiTest extends RestApiTest {
   }
 
   @Test
-  @Config(confParams = {ConstantPropertyKey.Name.USER_LINEAGE_ENABLED, "true"})
+  @Config(confParams = {PropertyKey.Name.USER_LINEAGE_ENABLED, "true"})
   public void getLineageInfoList() throws Exception {
     AlluxioURI input = new AlluxioURI("/input");
     AlluxioURI output = new AlluxioURI("/output");
@@ -125,7 +125,7 @@ public final class LineageMasterClientRestApiTest extends RestApiTest {
   }
 
   @Test
-  @Config(confParams = {ConstantPropertyKey.Name.USER_LINEAGE_ENABLED, "true"})
+  @Config(confParams = {PropertyKey.Name.USER_LINEAGE_ENABLED, "true"})
   public void reinitializeFile() throws Exception {
     mLineageClient.createFile(new AlluxioURI("/test")).close();
 
@@ -141,7 +141,7 @@ public final class LineageMasterClientRestApiTest extends RestApiTest {
   }
 
   @Test
-  @Config(confParams = {ConstantPropertyKey.Name.USER_LINEAGE_ENABLED, "true"})
+  @Config(confParams = {PropertyKey.Name.USER_LINEAGE_ENABLED, "true"})
   public void reportLostFile() throws Exception {
     mLineageClient.createFile(new AlluxioURI("/test")).close();
