+++ tests/src/test/java/alluxio/master/lineage/LineageMasterIntegrationTest.java
@@ -12,7 +12,6 @@
 package alluxio.master.lineage;
 
 import alluxio.AlluxioURI;
-import alluxio.ConstantPropertyKey;
 import alluxio.Constants;
 import alluxio.IntegrationTestUtils;
 import alluxio.LocalAlluxioClusterResource;
@@ -145,7 +144,7 @@ public class LineageMasterIntegrationTest {
    */
   @Test(timeout = 100000)
   @LocalAlluxioClusterResource.Config(
-      confParams = {ConstantPropertyKey.Name.MASTER_LINEAGE_CHECKPOINT_INTERVAL_MS, "100000"})
+      confParams = {PropertyKey.Name.MASTER_LINEAGE_CHECKPOINT_INTERVAL_MS, "100000"})
   public void lineageRecovery() throws Exception {
     final File logFile = mFolder.newFile();
     // Delete the log file so that when it starts to exist we know that it was created by the
